requirements.txt

pyrogram
pillow
rembg
